feat: updated planos, inserted detailed about gamers and plays

+++ microservices/jogos/aluragames/src/main/kotlin/br/com/roberto/aluragames/utilitario/JogoExtension.kt
@@ -4,5 +4,5 @@ import br.com.roberto.aluragames.modelo.InfoJogoJson
 import br.com.roberto.aluragames.modelo.Jogo
 
 fun InfoJogoJson.criaJogo(): Jogo {
-    return Jogo(this.titulo, this.capa, this.preco, this.descricao)
+    return Jogo(this.titulo, this.capa, this.preco.toBigDecimal(), this.descricao)
 }
